Fix compilation warnings
Fix everything `-Wpedantic -Wall -Wextra` whine about.

+++ src/main.c
@@ -115,7 +115,7 @@ int main(int argc, char *argv[])
         }
         SDL_RenderCopy(rend, co2_present ? present : not_present, NULL, &textrect);
 
-        for(int i = 0; i < sizeof(handled_scancodes); i++) {
+        for(size_t i = 0; i < sizeof(handled_scancodes); i++) {
             uint8_t handled_scancode = handled_scancodes[i];
             if(keys[handled_scancode] != prev_keys[handled_scancode]) {
                 net_send_keycode(remote, net_encode_scancode(handled_scancode, keys[handled_scancode]));
